requirements.txt
numpy
streamlit
pandas
json
os
openai
plotly.express